+++ org.knime.dl/src/org/knime/dl/core/data/DLDefaultIntBuffer.java
@@ -176,7 +176,7 @@ public class DLDefaultIntBuffer extends DLAbstractWrappingDataBuffer<int[]>
 	}
 
 	@Override
-	public void zeroPad(final int length) throws IllegalArgumentException, BufferOverflowException {
+	public void zeroPad(final long length) throws IllegalArgumentException, BufferOverflowException {
 		checkArgument(length > 0);
 		checkOverflow(m_nextWrite + length <= m_capacity);
 		for (int i = 0; i < length; i++) {
