+++ org.knime.dl/src/org/knime/dl/core/data/DLDefaultLongBuffer.java
@@ -160,7 +160,7 @@ public class DLDefaultLongBuffer extends DLAbstractWrappingDataBuffer<long[]>
 	}
 
 	@Override
-	public void zeroPad(final int length) throws IllegalArgumentException, BufferOverflowException {
+	public void zeroPad(final long length) throws IllegalArgumentException, BufferOverflowException {
 		checkArgument(length > 0);
 		checkOverflow(m_nextWrite + length <= m_capacity);
 		for (int i = 0; i < length; i++) {
